git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@7934 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ cell_based/src/odes/TysonNovak2001OdeSystem.hpp
@@ -76,7 +76,7 @@ private:
     /** Dimensionless parameter J_5. */
     double mJ5;
     /** Dimensionless parameter n. */
-    double mN;
+    unsigned mN;
     /** Dimensional parameter k_7. */
     double mK7;
     /** Dimensional parameter k_8. */
